pythonPackages.jq: 1.5.0 -> 1.6.0
The `jq` Python package stopped building in v. 1.5.0 due
to a failing unit test. I assume the reason to be that
it is currently built against a different version (1.7)
of `jq` than this version was originally written against
(at least, the test doesn't fail if I remove the patch
doing that).

Upgrading the package fixes this issue.